created helper_function.py
